refactor(measurement): split case analysis and analysis I/O
